nDisplay "Frustum Split": Tiled rendering of viewports that can be used in MediaIO.
- Added a new feature: splitting a viewport into multiple tiles for rendering as separate viewports.
  * Added tile customization for ICVFX camera.
  * Create/update/delete tiled viewports at runtime (MU/BP).
  * Added custom overscan for tiled viewports. Optimization by tile position.
  * Modified 'link' projection policy to support frustum projection splitting for tiled viewports.

- Added changes to the DisplayClusterConfiguration module:
  * Basic tile settings in the 'DisplayClusterConfigurationTypes_Tile.h' file.
  * Tile settings for ICVFX camera.
  * Tile settings for media configuration in the ICVFX camera.

- Added new callback OnDisplayClusterUpdateViewportMediaState() to the DisplayCluster module.

- DisplayClusterMedia module has been redesigned.
  * Added support for ICVFX camera tile viewports.

- Cleans up code in the DisplayCluster module.
  * The implementation of viewport and viewportproxy rules has been changed.
  * Functions were moved to separate files for FDisplayClusterViewport and FDisplayClusterViewportProxy classes.

#jira UE-198907
#rb Alejandro.Arango, Vitalii.Boiko, Andrey.Yamashev

[CL 30663336 by peter tarasenko in ue5-main branch]

+++ Engine/Plugins/Runtime/nDisplay/Source/DisplayCluster/Private/Render/Viewport/Containers/DisplayClusterViewport_CustomFrustumRuntimeSettings.h
@@ -26,17 +26,17 @@ struct FDisplayClusterViewport_CustomFrustumRuntimeSettings
 	* @param InOutTop                  - the value of the top projection plane that you want to change
 	* @parma InOutBottom               - the value of the bottom projection plane that you want to change
 	*/
-	static bool UpdateProjectionAngles(const FDisplayClusterViewport_CustomFrustumRuntimeSettings& InRuntimeSettings, float& InOutLeft, float& InOutRight, float& InOutTop, float& InOutBottom);
+	static bool UpdateProjectionAngles(const FDisplayClusterViewport_CustomFrustumRuntimeSettings& InRuntimeSettings, double& InOutLeft, double& InOutRight, double& InOutTop, double& InOutBottom);
 
 	/**
 	* Values in percent
 	*/
 	struct FCustomFrustumPercent
 	{
-		float Left = 0;
-		float Right = 0;
-		float Top = 0;
-		float Bottom = 0;
+		double Left = 0;
+		double Right = 0;
+		double Top = 0;
+		double Bottom = 0;
 	};
 
 	/**
